+++ example/person/openAccount.js
@@ -1,6 +1,5 @@
 /**
  * 口座開設サンプル
- * @ignore
  */
 const moment = require('moment');
 const open = require('open');