+++ example/person/searchOwnershipInfos.js
@@ -1,6 +1,5 @@
 /**
  * 所有権検索サンプル
- * @ignore
  */
 const moment = require('moment');
 const open = require('open');